+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Fix a false positive for `RSpec/LetSetup` when `let!` used in outer scope. ([@ydah])
 - Fix a false positive for `RSpec/ReceiveNever` cop when `allow(...).to receive(...).never`. ([@ydah])
 - Fix detection of nameless doubles with methods in `RSpec/VerifiedDoubles`. ([@ushi-as])
+- Add `AllowedIdentifiers` configuration option to `RSpec/RedundantPredicateMatcher`. ([@ydah])
 
 ## 3.7.0 (2025-09-01)